feat: backend sends string field information
Hack the database and add string field ids for an artidoc document in
the table plugin_artidoc_document_tracker_field.

Every REST representation of an artifact section has now a `fields`
entry. If fields configured in the document belongs to the tracker of
the artifact, then the value is rendered in the representation if the
user is allowed to read the field and the field is a string.

Part of story #42242: Artidoc read-only fields: string

Change-Id: I3ba29d124246a0161582c83c238461a9da033dec