fix: request #42262 PHP fatal error when processing async notifications
How to test:
- Given a tracker that is child of another in the hierarchy (for
example Task < User Story)
- When you create or edit a Task artifact, there should be no PHP
fatal error in the worker_syslog while sending e-mail notifications.
- Access to trackers in special project cases should still be denied in
Cross-tracker search. For example, you still cannot query trackers in
deleted projects with Cross-tracker search.

Why?
URLVerification is not suitable to use here because this check is only
valid for web requests, but TrackersPermissionsRetriever is meant to be
used also in REST requests and async processing. It should not depend on
global variables.

Change-Id: I9e596b8772bc23000ea456761842febfadc253cc