fix: request #42571 Cannot delete a freshly created Git repository
Introduced by 798eadc52b4e1bf30993cda794725164bbedc0d9.

We are now deleting the repository folder using the `gitolite` user
so we will now have an issue with permissions anymore.

In your dev environment to test this patch you will have to deploy the
sudoers file manually. In production it will be deployed by the
packaging.

Change-Id: Ic834c68617a0e3898143b8fa1d0352f47ca6a32b