feat: Style conflict in create artifact link
Part of story #41466 Have new Artifact Links in Artifact View

- * of required inputs
- vertical align of elements in tracker select
- vertical align of link type select

Note: I used the same style than in artifact modal. The color circle of
tracker should have a height 20px. It fixex the alignement with the label.

Others list fields of the view are impacted and their color circles are
now smaller (but it corresponds to the size in modal 🤷)

I'm not very happy with the fix for the link type but I do not find
another way to do it

*Testing:*

There should be no more weird style for artifact creation in art link
field

Change-Id: I2f639555bfbd3f43847729dcd56cc7a3aa556cda