chore: request #42578 Refresh dev documentation to clarify how features are developed
The documentation is rewritten to focus on the reality of contribution
of Tuleap: 99.9% of contributions comes from Enalean. There is no need
to describe a non-existing scenario where entire features are developed
externaly.

This commit also cover the "Artifact" part of the development process
and how the features are designed, especially with security and privacy
in mind.

Change-Id: I3d7eb5e57a3f572df8c0933e177116176e7f8562