feat: be able to click on the field selection tab
Part of story #42242: Artidoc read-only fields: string

This commit only allows to click on the field selection tab if a tracker is selected.
It should not display anything in its body for now.

Note:
The `<span>` have been replaced by buttons to be more accessible: the cursor is a pointer by default and it is possible to switch from one button to another using the keyboard.

At the moment, the button styles are overridden. A dedicated patch will define the navigation buttons in TLP, and this part will be updated.

How to test:
-- Hack plugin_artidoc_document_tracker table and put 0 as tracker id of your artidoc
-- Click on 'Configure document'
-> The Fields Selection tab should be disabled
-> A tooltip should be displayed on mouseover

-- Choose a tracker and save
-> The Fields Selection tab should now be clickable.

-- Try to navigate between the tabs through the keyboard
-> It should work

Change-Id: I7027b0e609e351ddff092e93e51261458b97d616