fix: Unable to change text field & comment formats
part of story #41466 Have new Artifact Links in Artifact View

- When you create a new artifact, you should be able to change the
format of text fields.
- If you choose "HTML" as default text format in your user preferences,
the CKEditors should be created at page load.
- When you edit an existing artifact, you should be able to change the
format for the new comment and for text fields.
- If you choose "HTML" as default text format in your user preferences,
the CKEditor for the new comment should be created at page load. When
you type into this CKEditor, the "Submission bar" with the submit
buttons should appear.
- No other functional change expected:
  - List fields with field dependencies still work.
  - When you type into CKEditors, the "Submission bar" should appear as
before.
  - You can still click on the "pencil" icons to edit fields
  - Required fields should work as before.

Change-Id: I7b20b558df79ee3424efd05dbf080d230be26a46