feat: Fields can be configured via REST
part of story #42242: Artidoc read-only fields: string

User can use REST API to configure the document to use a given set of
fields:
PUT artidoc/:id/configuration
{
  "selected_tracker_ids": [
    186
  ],
  "fields": [
    {
      "field_id": 62293,
      "display_type": "column"
    }
  ]
}

The submitted fields must be string field, readable, used, not in title
or description semantics, and belong to the selected tracker.
The fields are now correctly saved in database.

Change-Id: I47f5fe093df00ad966c97e102dde56f66046b0da