feat: Configuration route validates fields
part of story #42242: Artidoc read-only fields: string

How to test:
Use the REST API to configure an artidoc to display a set of fields:
PUT artidoc/:id/configuration
{
  "selected_tracker_ids": [186],
  "fields": [
    {
      "field_id": 62293,
      "display_type": "column"
    }
  ]
}

The submitted fields must be string field, readable, used, not in title
or description semantics, and must belong to the selected tracker.
⚠️ The fields are not yet saved to database to ease (a little) the
review. Only validation that the fields are correct is done.

Change-Id: I15bf9d40e3f30b1fbd996aa5f0046fbc71e6a11c