feat: request #42586 Do not always log business errors in Artidoc
How to test:
No change on the REST API responses for Artidoc REST routes.
None of the Faults are logged to the REST API syslog anymore, except
when Tuleap cannot write a file while uploading an image. This means
either the disk is full, or there is some permission or IO issue that
needs to be investigated by a human. This is the only case where the
fault is not a "business" error.

Note:
This error could (should ?) be represented as a regular Exception,
because the operations team managing Tuleap must be alerted, and it's
not an "expected" situation. All other cases are expected, normal
operation of Tuleap.

Change-Id: I8e9c41d4db3b09377f807426fcdeeddf52fe6666