chore: request #42588 Remove unneeded pnpm overrides
No functional change. CI should be happy.

Why?

Some overrides for pnpm do not actually need to stay. Overrides are
needed when a transitive dependency has a fixed range of versions for a
package, and we need to force it to use a later version (often for
security reasons). When the range is not fixed (for example with "*"),
we do not need to use overrides.

I also took the opportunity to refresh the nolyfill overrides to speed
up "pnpm install".

Change-Id: I0e24ce6e785c4e16406f28f5e3d728ba6e404407