feat: Display new artifact link field in artifact creation
Part of story #41466 Have new Artifact Links in Artifact View

The new view for artifact link field is now available in artifact
creation view. The component does nothing and all changes are not sent.

Note: also fix the label font-size. It uses the one from .tlp-label
which is smaller than the flamming parrot one

*Testing:*

In a tracker with the feature enabled, create a new artifact. In the
creation form the artifact link field should use the new display and the
component should works.

Change-Id: Icbc0e698d1b74492b4c03e4ff83cab35182e7b35