feat: XML export/import of artifact link field configuration
Part of story #41466 Have new Artifact Links in Artifact View

To allow field to export its configuration, the class must have the
property `$default_properties` filled. So I need to override field
administration display to not render the property.

*Testing:*

Export a tracker with an artifact link field. If you import it the
corresponding field should have the same configuration

Change-Id: Ie5dbcf03656675dd891fabc077c440e5550e84cf