feat: Display link field fault
Part of story #41466 Have new Artifact Links in Artifact View

LinkField dispatch a 'WillNotifyFault' event, we now catch it and
display its message in an alert div just above the field.

To force the behaviour, you can comment lines 99-101 in
ArtifactLinkSelectorAutoCompleter.ts and then use the search input with
meilisearch disabled. It should display a fault 'Not found'.

*Testing:*

The alert div should works on edition and creation

Change-Id: Ie8569921848245e154c1bdb9e62458c4997b3800