refactor: split configuration modal in several components
part of story #42242: Artidoc read-only fields: string

With the introduction of readonly fields, the configuration modal has
now two tabs:
- one to configure the tracker
- one to select the readonly fields to be displayed below artifact
  sections

The issue here is that we have to manage two configuration UIs inside
the same modal. The current component architecture makes it tricky to
save whether a tracker or fields depending on which tab is currently
active.

The proposition is the following:
- tabs have their own component
- tracker configuration UI has its own component
- fields configuration UI has its own component
- modal footer with the save/cancel buttons has its own component
- the right configuration UI is displayed depending of the currently
  active tab
- each configuration UI holds an instance of the modal footer component.
  It receives a different onSaveCallback following its parent element.

How to test:
-- no functional change expected:
--> No tabs in the modal when the feature flag is off
--> Fields selection tab disabled when no tracker has been selected yet
--> You can switch tabs
--> You can change the current tracker and save it
--> When you try to add an existing artifact section to the document
while no tracker is configured, the config modal still opens, and the
add exisitng modal still opens after the tracker selection, allowing you
to search artifacts in the previously selected tracker.

Change-Id: I1b5032c9bcd9dca14e7c4795a66961337dc5b053