feat: LinkField dispatch DidChangeLinkFieldValue
Part of story #41466 Have new Artifact Links in Artifact View

Each time LinkField has a change, it dispatch the event
DidChangeLinkFieldValue with the field_id.

This event is caught by the artifact view and then writes the links
representation in a hidden input.

This hidden input has an empty initial value:

    {"field_id":<number>,"all_links":[]}

It will be initialized with the correct value in a future contribution

*Testing:*

Do some changes to the field, the value of the hidden input (located
just below the component) should change corresponding the the component
status.

If you submit the form, you should get a php fatal error "TypeError": it
is normal because the input value is not yet treated

Behaviour of old component and artifact modal should not be impacted

Change-Id: I1ee860a96ed168076c2589d2a57f9765f62d94df