fix: e2e tracker report tests for advanced selectbox selection
Part of request #42184 Add end-to-end tests to tracker report

It seems that when the “advanced toggle” button is clicked, a POST request is triggered.
So it's possible that Cypress is still seeing the old selectbox when the multiple selection is made.
I added `cy.wait()` before selections that use the advanced selectbox, to wait for the POST request to finish before selecting.

make tests-e2e should pass and the nightly should be happy.

Error:
```
`cy.select()` was called with an array of arguments but does not have a `multiple` attribute set.
```

Change-Id: Ia04809dc92dceac81847ded9e581909a6ffbc06e