feat: add description and ArtifactLink to Activities
This default template is missing Description Semantic and should have
ArtifactLink field.

ArtifactLink was not there initially because it was meant to be a
"simple" tracker and cross reference was preferred because it displays
all kind of cross references.

Now, ArtifactLinks are one of the most powerful feature of Tuleap and
all the new features are baking them as a first class citizen. The
default template should make use of that.

closes request #42609: Default "Activities" tracker template should have description and ArtifactLinks

Change-Id: Ibeed9c2ed31eb8d1efc202d610b8b54b4ebff1f5