feat: Display LinkField in artifact view submit form
Part of story #41466 Have new Artifact Links in Artifact View

After submitting the artifact view in case of error, a view is
displayed with all fields in edit mode.

This view remove all elements that are not for field edition, so the
mount point of LinkField element was also removed as it doesn't pass the
selector.

By adding class `tracker_formelement_read_and_edit` to the mount point
the field is now displayed. This class seems to do nothing and it
doesn't really matter as the component replace it.

*Testing:*

You can force the artifact view error in `ArtifactLinkValidator` but
early returning false. The LinkField should be displayed without errors.

Change-Id: Icd6416e94aaccc13d0fc8cdb05799ea87bf7b245