chore: vue < 3.5 -> 3.4.38
Implements request #42616 vue: 3.3.4, 3.4.5, 3.4.18, 3.4.19, 3.4.27,
3.4.30 -> 3.4.38

No functional change expected. All of the upgraded apps should function
as before.

Why?
A few changes have been made in provide because Typechecking now raises
an error.
Same for changes of functions in templates.
v-model on components is not supported before v3.5 so it has been
converted to classic v-bind/v-on in Git permissions per group.

Change-Id: I2235bb137f7b2ad239f20f2f4b9e2150553a1de8