feat: Add XTS to ALM, ART, Team, Portfolio templates
Here is the new layout of the dashboard of each template:

Full ALM
--------

```
    Dashboard
        Column Left (small)
            Project Milestones
            Heartbeat
        Column right (large)
            XTS
                All open artifacts (in self)
                All artifacts assigned to me (in self)
```

SAFe Agile Release Train
------------------------

```
    Dashboard
        Roadmap
        Kanban
    Full Train view
        XTS
            All open artifacts in (self, aggregated)
            All artifacts assigned to me in (self, aggregated)
```

SAFe Scrum Team
---------------

```
    Dashboard
        Column Left (small)
            Project Milestones
        Column right (large)
            XTS
                All open artifacts (in self)
                All artifacts assigned to me (in self)
```

SAFe Portfolio
--------------

```
    Dashboard
        Column Left (small)
            Note to describe Portfolio
        Column Right (large)
            All open artifacts (in self)
            All artifacts assigned to me (in self)
```

When XTS is not enabled on the platform then the widget is not added to
the dashboard. And if it was the only widget of the dashboard (for
example Full Train View in Agile Release Train template) then the
dashboard is not added to the project (we don't add empty columns and
empty lines, we should do the same for dashboards).

Part of story #42604: Update default templates with Cross Tracker Search

Change-Id: I7a9fa57195906887bbca00755cf435775d8863ba