refactor: request #44479 Do not bundle node-gettext with all Vue apps
How to test:
No functional change. Vue apps are still translated as before.
Libs using node-gettext (for example list-picker) are still translated
as before.
Apps displaying formatted dates function as before, the dates should be
formatted according to user language (03/09/2025 or 2025-09-03).

Why?
vue3-gettext-init should not depend on @tuleap/gettext, because it pulls
node-gettext as a side-effect. To deal with shared code, we introduce a
dedicated lib @tuleap/locale that knows how to read the right
data-attribute and has the locale constants and the utility functions
getPOFileFromLocale + withoutExtension. This way, both translation
systems can depend on @tuleap/locale and Vue apps weight is reduced.

For example: tracker-creation: 240.45 kB -> 206.92 kB

We also reuse it for @tuleap/date-helper that needs the locale to
properly format dates. Instead of throwing an error, we fall back to the
default "en_US" locale.

Change-Id: I765ad0dc5dba1cbe039719db3f8d885f07bc19ce