fix: Flaky tracker report e2e test
part of request #42184 Add end-to-end tests to tracker report

How to test:
- Edit Tracker_FormElement_Field_List::fetchCriteriaValue()
- Add the following line at the beginning to add 300ms artificial
slowness to the endpoint that returns HTML for the advanced criteria

usleep(300 * 1000);

- make tests-e2e should succeed for tracker_report.cy.ts

Why?
I extracted a function to abstract the querying of a criterion block
(the "li"). It allows us to repeat the whole querying after waiting for
the async HTML. If we keep a reference to the old block, it causes
issues because Cypress finds a selectbox without "multiple" attribute
(which is only present in the advanced criteria). By making Cypress
query again the whole criterion block, it should find the "new" HTML for
the advanced criteria.
Having a function also removes some calls to cy.within(), which changes
the scope of the cy.get() and can be hard to deal with.

Change-Id: I5ffe1319e5432bfef62b47996338542096785555