feat: change the modal configuration workflow
part of story #42242: Artidoc read-only fields: string

Until now, the modal workflow was as follows:
Open the modal > Select a new tracker > Save -> a Close button appears.
Even if you change tabs to go to field selection, the Close button is still there and  it is not possible to save selected fields. You have to close the modal and then reopen it to select the fields.
This workflow makes the modal a little complex to use.

To make it easier to use it, I suggest the following workflow:
1. Each time you change tab, the save and cancel buttons reappear.
2. Saving the tracker and saving the fields is done separately. This means you can change the tracker, save it, then go to the select fields tab, and directly select the fields relating to the tracker just saved.

How to test:
-- Change the tracker and Save
-> A success message ans a Close button are displayed

-- Go to the Selected Field Tab
-> The just saved tracker is displayed in the info message
-> The Save and Cancel buttons are displayed in the footer
-> You can choose fields related to the tracker

-- Click on the Selected Tracker Tab
-> The Save and Cancel buttons are displayed
-> You can choose a tracker again

Change-Id: I48553173ccd75aabc2f94e054dd7070c099a6cd2