test: request #42636 Program management end-to-end test is flaky
No functional change. make tests-e2e should pass for
program_management.cy.ts

Why?
Verifying that the team we just linked is present in the list of teams
should force cypress to wait for the page reload after the PUT request
to add the team.
Timestamp in seconds should be enough for non-collision, so we drop the
last 4 digits.

Change-Id: I14d86fcd85d6d49194923beab61c1fcb9b67910b