fix: stop displaying selected fields in a tlp-table
part of story #42242: Artidoc read-only fields: string

Why?

The drag and drop experience is bad because the fields are displayed
in a table. Its height being computed by summing the heights of the rows it
contains, it produces a lot of flickering when the user tries to drop an
item at the wanted place.

How to test:
- reorder the fields by drag and dropping them
--> It is far easier/smoother than before

Change-Id: I03a50f0809216ac0b5d0eede7413e5cc25bfee96