fix: Stop storing CSRF protection tokens into the PHP sessions
PHP sessions can be cleaned during upgrades and we are forced
to limit the number of valid tokens at one time to be sure to
not end up with gigantic sessions content. Those 2 facts cause
usability issues where valid requests are rejected because the
token is not available anymore.

In order to prevent this contribution changes the way the challenges
are generated. We now attribute a random key to each user session and
we generate the challenge by signing this session key and the action
(e.g. form URL) with a private key known only by the server. This
make the CSRF tokens unique per session ID / form action and
unpredictable (as they currently are).

Performance should be roughly equivalent: while computing the
signatures require some compute time, the serialization and
deserialization of the PHP session is also not free.

This new approach has been retrofitted into the existing API to limit
the impacts accross the codebase.

Closes request #42590: Random issue when updating tracker tickets

Change-Id: I38ea460e1f051243406d6f114cb9a5f3dab64b4c