Taskbboard App is now in vue3
closes: request #41443 convert taskboard into vue3

How to test:
Check that you can still interact with taskboard (drag and drop,
create/edit card in places, edit effort/assignees).
No functional changes.

Why?
CardAssignee component is in part re-written. In Vue 3 it stopped
working properly, and we can make it simpler because we are no longer
limited to a single root element.

Co-authored-by: Marie Ange Garnier <marie-ange.garnier@enalean.com>
Co-authored-by: Joris Masson <joris.masson@enalean.com>
Change-Id: Ia5c07dacb01506c99ae219e96dfee8c252e1f972