chore: Drop Mockery usage in Workflow/PostAction/Update
Note: the removal of TransitionFactory was mandatory because there was a
usage of Mockery's `byDefault()` so we had to inline the methods of
TransitionFactory. This forced us to touch every files in
PostAction/Update/Internal, sorry for the long review.

Part of request #21775 Favor PHPUnit mock system over Mockery - part 2

*No functional changes expected*

Change-Id: I6c2f3a666e55706e22cbaf354ba45d4213524a35