chore: Drop Mockery usage in Workflow/PostAction
Note: the removal of TransitionFactory was mandatory because there was a
usage of Mockery's `byDefault()` so we had to inline the methods of
TransitionFactory. This forced us to touch every files in
PostAction/Update/Internal, sorry for the long review.

Part of request #21775 Favor PHPUnit mock system over Mockery - part 2

*No functional changes expected*

Change-Id: I147c359ea2e653a9caf5c3d2fdd1ffdd951f2669