fix: request #42957 Fields with no values are not filled when creating artifact from API
In `FieldsDataBuilder` we iterate on submitted values to fill an array
of values to save. If some fields are not submitted then they have no
value to save.

To fix that, it now "just" iterate on all tracker fields and if some
have no values it sets the default one. It skips all readonly fields and
ones user cannot submit/update

*Testing:*

Have a tracker with 2 non-mandatory fields. Create a new artifact using
api and give a value for only one of those fields. Then get this
artifact also from API:

Before -> only one field value is returned
After  -> both fields are returned with the default value for the second
          one

Change-Id: I4ae3c4bd388f84d956e5509446ee2c2274cfe1ae