fix: request #42959 The contents of a step disappear
When editing a test case, if you add a step and create a new step before it, the content of the previously created step is deleted.

How to test:
-- Go to a test case that contains at least 2 steps
-- Click on the button to add a step between the 2 and enter content
-- Click on the button to add a step above it
=> The content of the step should always be displayed

-- Click on the Preview button
=> Content should be displayed in Markdown

This problem seems to have been introduced by this patch:
cf21a05210faf6711c4c667a52010744da812f96

Why?
It seems that when we switched to compat-mode, the props that were mutated were transformed into local refs.
But in this way, the changes to these properties are not applied to the other components.
To correct this, I added events to update the parent component's property, so that changes are shared with other components.

Change-Id: I65cf378463f2dfaeea718812f52cd4c09a76d9e9