fix: request #42968 Fatal error in cardwall report with user open list
In Tracker_FormElement_Field_List_Bind_Users::getValueFromRow it creates
a new Tracker_FormElement_Field_List_Bind_UsersValue. The constructor
asks for a UUID as first parameter but it was not given.

*Testing:*

Have a tracker with an open list bind to users. You should be able to
use the carwall view of tracker report

Change-Id: I380ab4c75c4cc0aacea824ecce7e0dca7335f66f