feat: Create artifact with new artifact link field
Part of story #41466 Have new Artifact Links in Artifact View

CreateArtifactAction now uses the artifact link field value to create
forward and reverse links.

Note: I have extracted the code that generates the
ChangesetValuesContainer as it is the same for update and create
artifacts.

*Testing:*

You should be able to create artifacts linked (or not) to other artifacts

Change-Id: I1775ea527a4cd82daa4ab2ef524f33e64de48be5