fix: request #42975 Program errors are not always shown in administration
When building the program admin presenter if there is no tracker found
for iterations it throws an error and check on increment is not done. So
if there is misconfiguration for increment it is not displayed.

Plus the errors even collected are displayed only if there is errors for
increment and iteration and planification. If one of them doesn't have
an error nothing is displayed

*Testing:*

Have a program configuration with increments but not iterations. In a
team project remove the description semantic for the milestone tracker.

You should get an error in program administration.

Change-Id: I3ae35f502a5a40351b2eea4857294d7ea22c9926