test: add artifact link field when submitting all fields
Part of story #41466 Have new Artifact Links in Artifact View

When we created the test that submits an artifact with all fields, we left out the artifact link field as it was under development.
Now that it is finished, we can add this field.

How to test:
make tests-e2e should pass

Change-Id: I021deb7970c1c27167adf971f14106e1deec5e41