fix: request #42974: Do not keep passwords of users authenticating with an OIDC provider
Tuleap is now making sure no password is stored for users authenticating
with an OIDC provider.

A trade-off between usability and security is made when a user with a
link to an OIDC provider is trying to use the password-based auth.
Tuleap now indicates that the account is linked to an OIDC provider and
that the user should use that to login instead of the generic error
message. This leaks the presence of the account on the instance but it
avoids confusion for the end-users and support team. The same trade-off
is already done for the passwordless auth.

Change-Id: Idd2116c125a0c8c58745e274d8aaa6712865cec5