chore: request #42977 Harden CSRF protection by blocking all cross-site state modifying queries to the Web UI
No functional changes for legitimate requests, for others requests they
will now get an early 403.

An easy way to test is to have a local looking like this:
```
<!DOCTYPE html>
<html>
  <body>
    <form method="post" action="https://tuleap-web.tuleap-aio-dev.docker/account/login.php">
      <button type="submit">OK</button>
    </form>
  </body>
</html>
```

Before this change the request would be rejected by CSRF token
mechanism, now you get an early 403 as a defense in depth approach.

As Tuleap serves everything under the same origin, requests with a
`Sec-Fetch-Site` header set to `same-site` are now considered in the
same way than cross-origin requests.

Change-Id: I80e7bd0e977320d2bd34bc4cb4f8238b57491c9a