fix: Set existing art link fields to old view
Part of story #41466 Have new Artifact Links in Artifact View

When checking if the field should use the new view, we get its
preference value in db. If no value is present (the case for almost all
fields) then the default value is used -> new view.

It should not be the case -> existing fields should stay with the old
view

*Testing:*

Existing fields should use the old view

Change-Id: I231c322630d521e156cc9fc6d7e9730598b32601