fix: Artifact Links view should use old view
Part of story #41466 Have new Artifact Links in Artifact View

It was using new view but as the component script is not included
nothing was rendered. Instead it should use the old view.

*Testing:*

In a tracker using the new field, links tab of artifact view should
render the old table.

Change-Id: I08108b960618af454336b08cae66b0ffa828296b