fix: Impossible to click on a card behind the add-in-place button
part of request #42956 Style kanban scrollbars using CSS only

How to test:
- Set up a kanban with one column that is full, with more cards than
there is space vertically for the column.
- Scroll down to the bottom of the column.
- You should be able to click on the last card on the bottom, to
expand/collapse it, edit it with the modal, etc.
- The add-in-place "+" button can now be focused through the keyboard
"Tab" key.
- No functional change expected for the add-in-place function: you can
still open it using the button, you can still close it with "Esc" key,
you can still create cards with it.

Why?
Removed some CSS classes that were redundant or not needed. Renamed the
remaining with "kanban-add-in-place" prefix.

Change-Id: I75c34356890756ddefe681ab8ffd1a05e15972b6