Revert "fix: request #42957 Fields with no values are not filled when creating artifact from API"
This reverts commit 86b78e9a1a52f065eb8ebfc05ff9dcb6bac50f31.

It has introduced bugs when creating artifacts with just their title ->
at least, the status field is not filled with the default value but none
and does not appear in corresponding kanban.

Need further investigations to check what is the exact problem in this
patch but with the near release it is safer to revert it for now.

Change-Id: I2d9f277f89970af36b51f7ccee4b1a9ecfc9d168