refactor: DRY setup for Section transformer test
part of story #42248 Artidoc read-only fields: User Groups list fields

No functional change. Unit-tests should pass.

Why?
The user is always the same, so we might as well make it a property
instead of a parameter.
An interface was mocked instead of using the stub.
A very big part of the setup is always copy/pasted, which makes the test
much longer and harder to understand, it's hard to find the differences
between tests in 100 lines of setup.

Change-Id: I160f23c2a50d115286c6b194f7027819312e0ee0