chore: Replace "successive" methods by more realistic stub behaviour
part of story #41466 Have new Artifact Links in Artifact View

No functional change. CI should be happy.

Why?
We used to write stubs with "successive" methods, but it's dangerous as
it makes the test setup dependant on the order of method calls. Stubs
can be closer to reality by checking the ID they receive and choosing to
return something only if they have been properly prepared. It also
forces us to be more precise in our setup and explain which tracker the
fields and artifacts come from, which is not a bad thing.

Change-Id: Ie7606eb4ad1ff5afb64d00125171c4093fe806ff