fix: request #42988 Program increments are not synchronized when adding a team
When a ProgramIncrement is built, the title of the program increment is
retrieved. If the title semantic is defined, and the field readable by
the current user, then the title is returned, else it is null.

However, the current user is never passed to Artifact::getTitle(),
meaning that when the worker in charge of the synchronization is run, a
fake user will be used (id = 0). Hence, the synchronization fails.

To prevent this behavior, we now retrieve the title value from the last
changeset of the field bound to the title semantic using the current
user making the request.

How to test:
- Have a program project with aggregated team projects and mirrored
  program increments
- Remove the readonly permission on the field bound to the title
  semantic to "all users" in the program increments tracker
- Add a new team project to the program project
- Click the [Synchronize] button

--> It works

Change-Id: Id0f529d5568e7b592fe420ce0db852567e0c0d98