feat: Translate errors for reverse links
part of story #41466 Have new Artifact Links in Artifact View

How to test:
- Using the REST API (for example the artifact modal), create a typed
reverse link (for example Child Of) to an artifact where there is no
Artifact Links field. You should get an I18N Rest exception with a
translated error message (unfortunately, it's not yet made visible in
the modal UI).
- Do the same in the artifact view. You should get an error feedback
with the translated error message.
- The other error can only be triggered by REST API if you try to set a
reverse link to an artifact that you cannot see or that does not exist.

Why?
Translation of error messages is always done at the top-most endpoint,
because it depends on the context of the action. We could make them
different between the UI and the REST API. Here I chose to use the same
error messages. This is also one reason why the messages in Faults are
not translated.

Change-Id: I444914ec3375b6ea63cb3988aaf7718d69b7eb72