chore: fix gitlab group link wizard gettext import
part of request #42615 vue-tsc: 1.8.11 -> 2.2.8

No functional change when you create a new project.

Why?
vue-tsc when upgraded to v2 raises errors for usage of $gettext in the
template. Proper usage is to import useGettext and then call it to get
access to $gettext.
It also stops raising errors if we "just" use VueWrapper. Since there is
nothing specific exposed in those components, it's alright.

Change-Id: I6ff39eabd6f39bff9aee526bdc9d80b27ed93f12