feat: Introduce a CLI command to unlink user from providers
Since request #42974, it is impossible to remove the last provider for a
user account. It is painful during the execution of Tuleap tests
campaigns, where it is required to add and remove providers to test the
feature.

Now it is possible to issue the following command:
tuleap oidc_client:remove-all-user-links <user_name>

=> the user won't have anymore OIDC providers linked to their account

Change-Id: I35db297e24434f89b3957b417edd5c9c6b30cc1d