refactor: DependencyNatureControl tests to simplify the bump to Vue 3
Part of request #41441 Migrate Roadmap widget to Vue 3

Why?
The tests are very repetitive and difficult to read, as there is a `shallowMount` in each test. To facilitate bump to Vue3, I think it makes sense to refactor some tests to add a `getWrapper()` method and avoid duplication.
Indeed, when switching to Vue 3, all `mocks` will have to be modified to transform them into store options by using the `getGlobalTestOptions` function.

Given that a bump to Vue3 is a painful thing to do and review, I think it's a good idea to do it now, in a dedicated commit.

No functional change expected.

Change-Id: Ic746ddd2bd893a7209d10a2e280a944e8e4336c0