refactor: BarPopover tests to simplify the bump to Vue 3
Part of request #41441 Migrate Roadmap widget to Vue 3

Why?
The tests are very repetitive and difficult to read, as there is a `shallowMount` in each test. To facilitate bump to Vue3, I think it makes sense to refactor some tests to add a `getWrapper()` method and avoid duplication.
Indeed, when switching to Vue 3, all `mocks` will have to be modified to transform them into store options by using the `getGlobalTestOptions` function.

No functional change expected.

Change-Id: I19153825c7c98be3a67d7fff6e402dc4b0816496