refactor: TaskBar tests to simplify the bump to Vue 3
Part of request #41441 Migrate Roadmap widget to Vue 3

Why?
The tests are very repetitive and difficult to read, as there is a `shallowMount` in each test. To facilitate bump to Vue3, I think it makes sense to refactor some tests to add a `getWrapper()` method and avoid duplication.
Indeed, when switching to Vue 3, all `mocks` will have to be modified to transform them into store options by using the `getGlobalTestOptions` function.

No functional change expected.

Change-Id: I3984830e077128013e0d88777e6bc7de99ac51e9