fix: request #43328 Broken artifact tooltip when linking sprint
While retrieving recently viewed artifacts, ones from backlog (rel,
sprints) have the planning link for their uri. So when getting the json
for the tooltip we fetch on this uri and get the full html.

Instead it should get the artifact view uri -> when building the object
containing new link data, we now proxy the uri building with
`/plugins/tracker/?aid={id}`

*Testing:*

Link an artifact to a sprint, the tooltip should not be broken

Change-Id: I33c8c66a15348e07eca753177d5d7e7f2e36c7a6