feat: Artidoc sections route returns user groups list fields
part of story #42248 Artidoc read-only fields: User Groups list fields

How to test:
- Forge your DB, insert a row in plugin_artidoc_document_tracker_field
table. Choose a list field bound to user groups.
- Set up an Artidoc with one or more Artifact sections.
- Set up your artifacts to have one, several or no value for the list
field.
- You should see the field displayed in the Artidoc, for each Artifact
section.
- You cannot choose the field in the configuration modal yet.

Why?
Switched the "FieldType" string to snake_case to be more consistent with
CrossTrackerSelectedType which has a similar role. We may not have the
same categories, but users of the REST API will expect that we at least
have the same casing.

Change-Id: If87e700e756253482db37bc8352cbd8731cf50f0