feat: be able to expand a row
Part of story #42989 Expand level N-1 artifact link in Cross Tracker Search

This commit only allows you to click on the carets. When clicked, a new line appears below with only the artifact edit icon.
The next commit will display the correct number of skeletons for each arifact link.

The management of background coloring according to odd and even lines will be handled in a dedicated commit.

How to test:
-- Have a query with some artifacts with links
-- Click on a caret
=> It should display an empty new line
-- Click again on the caret
=> It should hide the previously displayed line

Change-Id: Iaba4a22bb12ea0b774faf9c5335a57b948483017