chore: Replace @import by @use in LDAP plugin
Part of request #42982 Fixable Sass deprecations

No functional change expected. Go to project administration, choose a
custom user group, and click on the button to link it to a LDAP user
group. The styles in the modal should not be broken.

Why?
The spinner gif was the same as the one in Burning Parrot, so I removed
the duplicate.

Change-Id: I425b40c338536034495448b5cd3ffe0beb781a90