test: DB test FieldDao for basic Create Read Delete
part of request #43337 Migrate FieldDao to EasyDB

No functional change. CI should be happy.

Why?
The first step in migrating an old DAO is to DB test it. Given the size
of this DAO, it will be done in several steps. With this first step, I
try to test create, get by ID  and delete.

There is a big condition in the DAO that has to do with guessing the
shortname of the field, but I'm skeptical that it is actually reachable,
so I prefer not to put effort into testing it right now.

original_field_id is also intentionally always left null because it is
related to shared fields. Testing shared fields properly will require
dedicated efforts, a lot of methods in the DAO are actually meant only
for shared fields.

Change-Id: Ie74b8bc241869cd2758105d681a21eeacd8baf41