refactor: remove Codendi_Request::getTime() ♻️
There are several ways to retrieve the request time in Tuleap:

* `$request->getTime()`
* `$_SERVER['REQUEST_TIME']`
* `RequestTime::getTimestamp()`

For the sake of consistency, we should keep only one way.
`RequestTime::getTimestamp()` is the safest because it does not trust
that the global array `$_SERVER` contains `'REQUEST_TIME'` key.

This contribution removes all occurrence of `$request->getTime()`.

No functional changes.

Part of request #43339: Unify retrieval of request time

Change-Id: I5689d204c470b42391ece853fb8a53ea1a61b054