refactor: Move arrows documentation to tlp-doc 📝
Since arrows are (will be) used by two different apps (roadmap and XTS),
there is no reason to keep the documentation in roadmap.

Moving it to tlp-doc will have the benefit to better document them,
notably about the usage. This contribution started to enhance it, but it
deserve to be more complete later, when we will have a clearer view of
what are the impact of integration into XTS (no need to spend too much
time documenting functions that can be changed in the near future).

Run `pnpm --filter=tlp-doc run storybook` and you will have access to
the documentation under TLP » Visual assets » Arrows.

Part of story #42989: Expand level N-1 artifact link in Cross Tracker Search

Change-Id: Ie7b74c72e3e8252676225539d45bfcd6bac7bd13